Refactor: Rename 'vehiclesList' to 'vehicles' styles folder and update imports
