+++ src/styles/features/vehicles/SplideCarousel.scss
