+++ src/styles/features/vehicles/VehiclesList.scss
